+++ packages/theming/win32-theme/src/NativeModule/getThemingModule.native.ts
@@ -1,12 +1,15 @@
 import { OfficeThemingModule } from './officeThemingModule';
-import { NativeEventEmitter, NativeModules } from 'react-native';
+import { NativeEventEmitter } from 'react-native';
 import { fallbackGetPalette, fallbackOfficeModule } from './fallbackOfficeModule';
 
 declare module 'react-native' {
-  interface NativeModulesStatic {
-    Theming: OfficeThemingModule & EventSubscriptionVendor;
+  interface ITurboModuleRegistry {
+    get: (name: 'Theming') => OfficeThemingModule & EventSubscriptionVendor;
   }
+  const TurboModuleRegistry: ITurboModuleRegistry;
 }
+
+import { TurboModuleRegistry } from 'react-native';
 
 /**
  *  If we have a userAgent string, let's assume we're web debugging.  __DEV__ is for developer bundles.  Currently,
@@ -19,7 +22,7 @@ function disableGetPalette(): boolean {
 }
 
 export function getThemingModule(): [OfficeThemingModule, NativeEventEmitter | undefined] {
-  const module = NativeModules && NativeModules.Theming;
+  const module = TurboModuleRegistry.get('Theming');
   // if the native module exists return the module + an emitter for it
   if (module) {
     // mock getPalette if it should be disabled, otherwise return the module directly