+++ apps/ios/CHANGELOG.md
@@ -1,8 +1,72 @@
 # Change Log - @fluentui-react-native/tester-ios
 
-This log was last generated on Thu, 26 Aug 2021 21:02:08 GMT and should not be manually modified.
+This log was last generated on Tue, 14 Sep 2021 19:31:04 GMT and should not be manually modified.
 
 <!-- Start content -->
+
+## 0.0.1
+
+Tue, 14 Sep 2021 19:31:04 GMT
+
+### Patches
+
+- Bump @fluentui-react-native/tester to v0.43.2 (ruaraki@microsoft.com)
+
+## 0.0.1
+
+Mon, 13 Sep 2021 23:22:43 GMT
+
+### Patches
+
+- Bump @fluentui-react-native/tester to v0.43.1 (ruaraki@microsoft.com)
+
+## 0.0.1
+
+Fri, 10 Sep 2021 21:30:12 GMT
+
+### Patches
+
+- Bump @fluentui-react-native/tester to v0.43.0 (30809111+acoates-ms@users.noreply.github.com)
+
+## 0.0.1
+
+Fri, 10 Sep 2021 18:16:59 GMT
+
+### Patches
+
+- Bump @fluentui-react-native/tester to v0.42.2 (ruaraki@microsoft.com)
+
+## 0.0.1
+
+Thu, 09 Sep 2021 22:12:19 GMT
+
+### Patches
+
+- Bump @fluentui-react-native/tester to v0.42.1 (ruaraki@microsoft.com)
+
+## 0.0.1
+
+Thu, 09 Sep 2021 20:03:01 GMT
+
+### Patches
+
+- Bump @fluentui-react-native/tester to v0.42.0 (v.kozlova13@gmail.com)
+
+## 0.0.1
+
+Tue, 31 Aug 2021 18:48:34 GMT
+
+### Patches
+
+- Bump @fluentui-react-native/tester to v0.41.0 (chripear@microsoft.com)
+
+## 0.0.1
+
+Fri, 27 Aug 2021 15:48:49 GMT
+
+### Patches
+
+- Bump @fluentui-react-native/tester to v0.40.1 (krsiler@microsoft.com)
 
 ## 0.0.1