+++ packages/components/PersonaCoin/CHANGELOG.md
@@ -1,8 +1,40 @@
 # Change Log - @fluentui-react-native/persona-coin
 
-This log was last generated on Thu, 26 Aug 2021 21:02:08 GMT and should not be manually modified.
+This log was last generated on Mon, 13 Sep 2021 23:22:43 GMT and should not be manually modified.
 
 <!-- Start content -->
+
+## 0.9.17
+
+Mon, 13 Sep 2021 23:22:43 GMT
+
+### Patches
+
+- Bump @uifabricshared/foundation-compose to v1.10.15 (ruaraki@microsoft.com)
+
+## 0.9.16
+
+Fri, 10 Sep 2021 18:16:59 GMT
+
+### Patches
+
+- Bump @fluentui-react-native/theme-tokens to v0.7.1 (ruaraki@microsoft.com)
+
+## 0.9.15
+
+Thu, 09 Sep 2021 22:12:19 GMT
+
+### Patches
+
+- Bump @fluentui-react-native/theme-tokens to v0.7.0 (ruaraki@microsoft.com)
+
+## 0.9.14
+
+Thu, 09 Sep 2021 20:03:01 GMT
+
+### Patches
+
+- Bump @fluentui-react-native/theme-tokens to v0.6.1 (ruaraki@microsoft.com)
 
 ## 0.9.13