+++ packages/deprecated/theming-react-native/CHANGELOG.md
@@ -1,8 +1,40 @@
 # Change Log - @uifabricshared/theming-react-native
 
-This log was last generated on Thu, 26 Aug 2021 21:02:08 GMT and should not be manually modified.
+This log was last generated on Mon, 13 Sep 2021 23:22:43 GMT and should not be manually modified.
 
 <!-- Start content -->
+
+## 0.12.17
+
+Mon, 13 Sep 2021 23:22:43 GMT
+
+### Patches
+
+- Bump @fluentui-react-native/default-theme to v0.8.4 (ruaraki@microsoft.com)
+
+## 0.12.16
+
+Fri, 10 Sep 2021 18:16:59 GMT
+
+### Patches
+
+- Bump @fluentui-react-native/default-theme to v0.8.3 (ruaraki@microsoft.com)
+
+## 0.12.15
+
+Thu, 09 Sep 2021 22:12:19 GMT
+
+### Patches
+
+- Bump @fluentui-react-native/default-theme to v0.8.2 (ruaraki@microsoft.com)
+
+## 0.12.14
+
+Thu, 09 Sep 2021 20:03:01 GMT
+
+### Patches
+
+- Bump @fluentui-react-native/default-theme to v0.8.1 (ruaraki@microsoft.com)
 
 ## 0.12.13