+++ packages/experimental/Icon/CHANGELOG.md
@@ -1,8 +1,40 @@
 # Change Log - @fluentui-react-native/icon
 
-This log was last generated on Thu, 26 Aug 2021 21:02:08 GMT and should not be manually modified.
+This log was last generated on Mon, 13 Sep 2021 23:22:43 GMT and should not be manually modified.
 
 <!-- Start content -->
+
+## 0.8.21
+
+Mon, 13 Sep 2021 23:22:43 GMT
+
+### Patches
+
+- Bump @fluentui-react-native/theming-utils to v0.7.2 (ruaraki@microsoft.com)
+
+## 0.8.20
+
+Fri, 10 Sep 2021 18:16:59 GMT
+
+### Patches
+
+- Bump @fluentui-react-native/text to v0.10.16 (ruaraki@microsoft.com)
+
+## 0.8.19
+
+Thu, 09 Sep 2021 22:12:19 GMT
+
+### Patches
+
+- Bump @fluentui-react-native/text to v0.10.15 (ruaraki@microsoft.com)
+
+## 0.8.18
+
+Thu, 09 Sep 2021 20:03:01 GMT
+
+### Patches
+
+- Bump @fluentui-react-native/text to v0.10.14 (ruaraki@microsoft.com)
 
 ## 0.8.17