+++ packages/framework/theme/CHANGELOG.md
@@ -1,8 +1,16 @@
 # Change Log - @fluentui-react-native/theme
 
-This log was last generated on Thu, 26 Aug 2021 21:02:08 GMT and should not be manually modified.
+This log was last generated on Thu, 09 Sep 2021 20:03:01 GMT and should not be manually modified.
 
 <!-- Start content -->
+
+## 0.5.19
+
+Thu, 09 Sep 2021 20:03:01 GMT
+
+### Patches
+
+- Bump @fluentui-react-native/theme-types to v0.12.1 (ruaraki@microsoft.com)
 
 ## 0.5.18